+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## master (unreleased)
 
+* [#933](https://github.com/clojure-emacs/cider-nrepl/pull/933) Add the `cider-inspector-print-current-value` command to print the current value of the inspector.
+
 ## 0.55.0 (2025-04-10)
 
 * Bump `orchard` to [0.33.0](https://github.com/clojure-emacs/orchard/blob/master/CHANGELOG.md#0330-2025-04-08).